src/server/pyproject.toml
[project]
name = "server"
description = "None"
readme = "README.md"
version = "0.1.0"

requires-python = ">=3.13"

dependencies = [
    "django>=5.2.4",
    "django-htmx>=1.23.2",
    "django-vite>=3.1.0",
    "django-cotton>=0.2.0",
    "sqids>=0.5.2",
    "python-dotenv>=1.1.1",
    "django-widget-tweaks>=1.5.0",
    "django-unfold>=0.65.0",
    "django-ninja>=1.3.0",
    "email-validator>=2.0.0",
    "django-allauth>=65.3.0",
    "django-file-form>=4.0.0",
    "boto3>=1.35.0",
    "django-storages>=1.14.0",
    "djangorestframework>=3.16.1",
    "random-username>=1.0.2",
    "psycopg2-binary>=2.9.9",
    "dj-database-url>=2.2.0",
    "whitenoise[brotli]>=6.7.0",
    "django-tasks>=0.11.0",
    "django-pydantic-field>=0.5.1",
    "pgvector>=0.4.2",
    "uvicorn>=0.40.0",
    "uvicorn-worker>=0.4.0",
    "gunicorn>=23.0.0",
    "honcho>=2.0.0",
    "djlint>=1.36.4",
    "pillow>=12.2.0",
    "django-anymail[amazon-ses]>=14.0",
    "django-import-export>=4.4.0",
    "django-taggit>=6.1.0",
]


[tool.djlint]
ignore = ["D018","H006","H021"]
custom_html = "c-[\\w-]+"
custom_blocks="buttons,endbuttons,fragment,component_title,component_navinfo,component_alert_global__acting_for,partialdef,component_box_nexus_activated_service"
max_attribute_length=200
preserve_blank_lines=true
extend_exclude = "itou/templates/utils/widgets/duet_date_picker_widget.html"
format_css = true
format_js = true
